requirements.txt
python-dotenv
langchain
PyPDF2
tiktoken
unstructured
pinecone-client
pandas
numpy
